Add files via upload
The objective of this project is to predict whether the customer will leave
the telecom or not. The data was imbalance so for balancing the data, I
have used Random Over Sampler. In this project, I have used all the
Classification Algorithm and by looking at the Accuracy and Recall I have
found AdaBoost (Adaptive Boosting) Classifier Algorithm gives the best
result.